+++ simtong-application/src/main/kotlin/team/comit/simtong/domain/notification/usecase/SendNotificationUseCase.kt
@@ -1,23 +1,23 @@
 package team.comit.simtong.domain.notification.usecase
 
+import team.comit.simtong.domain.notification.dto.request.SendMulticastNotificationData
+import team.comit.simtong.domain.notification.dto.request.SendNotificationData
 import team.comit.simtong.domain.notification.model.Notification
 import team.comit.simtong.domain.notification.model.NotificationReceiver
-import team.comit.simtong.domain.notification.model.NotificationType
 import team.comit.simtong.domain.notification.spi.CommandNotificationPort
 import team.comit.simtong.domain.notification.spi.NotificationQueryUserPort
 import team.comit.simtong.domain.notification.spi.SendPushMessagePort
 import team.comit.simtong.domain.user.exception.UserExceptions
 import team.comit.simtong.global.annotation.UseCase
-import java.util.UUID
-import kotlin.collections.List
 
 /**
  *
  * 알림 전송을 담당하는 SendNotificationUseCase
  *
  * @author kimbeomjin
+ * @author Chokyunghyeon
  * @date 2022/12/30
- * @version 1.1.0
+ * @version 1.2.5
  **/
 @UseCase
 class SendNotificationUseCase(
@@ -26,43 +26,43 @@ class SendNotificationUseCase(
     private val queryUserPort: NotificationQueryUserPort
 ) {
 
-    fun execute(userId: UUID, title: String, content: String, type: String, identify: UUID?) {
-        val deviceToken = queryUserPort.queryDeviceTokenByUserId(userId)
+    fun execute(request: SendNotificationData) {
+        val deviceToken = queryUserPort.queryDeviceTokenByUserId(request.userId)
             ?: throw UserExceptions.NotFound("디바이스 토큰이 존재하지 않습니다.")
 
         val notification = commandNotificationPort.saveNotification(
             Notification(
-                title = title,
-                content = content,
-                type = NotificationType.valueOf(type),
-                identify = identify
+                title = request.title,
+                content = request.content,
+                type = request.type,
+                identify = request.identify
             )
         )
 
         commandNotificationPort.saveNotificationReceiver(
             NotificationReceiver(
-                userId = userId,
+                userId = request.userId,
                 notificationId = notification.id
             )
         )
 
-        sendPushMessagePort.sendMessage(deviceToken, title, content, NotificationType.valueOf(type), identify)
+        sendPushMessagePort.sendMessage(deviceToken, request.title, request.content, request.type, request.identify)
     }
 
-    fun execute(userIds: List<UUID>, title: String, content: String, type: String, identify: UUID?) {
-        val deviceTokens = queryUserPort.queryDeviceTokensByUserIds(userIds)
+    fun execute(request: SendMulticastNotificationData) {
+        val deviceTokens = queryUserPort.queryDeviceTokensByUserIds(request.userIds)
 
         val notification = commandNotificationPort.saveNotification(
             Notification(
-                title = title,
-                content = content,
-                type = NotificationType.valueOf(type),
-                identify = identify
+                title = request.title,
+                content = request.content,
+                type = request.type,
+                identify = request.identify
             )
         )
 
         commandNotificationPort.saveAllNotificationReceiver(
-            userIds.map {
+            request.userIds.map {
                 NotificationReceiver(
                     userId = it,
                     notificationId = notification.id
@@ -70,6 +70,6 @@ class SendNotificationUseCase(
             }
         )
 
-        sendPushMessagePort.sendMessage(deviceTokens, title, content, NotificationType.valueOf(type), identify)
+        sendPushMessagePort.sendMessage(deviceTokens, request.title, request.content, request.type, request.identify)
     }
 }
